pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "frevl"
version = "1.0.0"
description = "FrEVL: Frozen Embeddings for Efficient Vision-Language Understanding"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Emmanuelle Bourigault", email = "emmanuelle.bourigault@research.org"},
]
maintainers = [
    {name = "Emmanuelle Bourigault", email = "emmanuelle.bourigault@research.org"},
]
keywords = [
    "vision-language",
    "multimodal",
    "clip",
    "vqa",
    "visual-question-answering",
    "efficient-ml",
    "frozen-embeddings",
    "pytorch",
    "deep-learning",
    "computer-vision",
    "nlp",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "clip @ git+https://github.com/openai/CLIP.git",
    "transformers>=4.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
demo = [
    "gradio>=4.0.0",
    "streamlit>=1.28.0",
]
serve = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "redis>=5.0.0",
    "prometheus-client>=0.18.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "myst-parser>=2.0.0",
]
all = [
    "frevl[dev,demo,serve,docs]",
]

[project.urls]
Homepage = "https://github.com/EmmanuelleB985/FrEVL"
Documentation = "https://emmanelleb985.github.io/FrEVL/"
Repository = "https://github.com/EmmanuelleB985/FrEVL"
Issues = "https://github.com/EmmanuelleB985/FrEVL/issues"
Paper = "https://arxiv.org/pdf/2508.04469"

[project.scripts]
frevl-train = "frevl.cli:train"
frevl-evaluate = "frevl.cli:evaluate"
frevl-demo = "frevl.cli:demo"
frevl-serve = "frevl.cli:serve"
frevl-download = "frevl.cli:download"

[tool.setuptools]
packages = ["frevl"]
include-package-data = true

[tool.setuptools.package-data]
frevl = [
    "configs/*.yaml",
    "configs/*.json",
    "assets/*",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
    "integration: marks integration tests",
]

[tool.coverage.run]
source = ["frevl"]
omit = [
    "*/tests/*",
    "*/benchmarks/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "pass",
    "except ImportError:",
]

[tool.pylint]
max-line-length = 100
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "W0622",  # redefined-builtin
]

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.bandit]
exclude_dirs = ["tests", "benchmarks"]
skips = ["B101", "B601"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.0.0"
tag_format = "v$version"
version_files = [
    "frevl/__init__.py:__version__",
    "pyproject.toml:version",
]
